requirements.txt
torch==1.0.0 
torchvision==0.4.0
scikit-image
tensorboardX
path.py
